--- src/main/java/Friday/parser/Parser.java
@@ -1,7 +1,0 @@
-package Friday.parser;
-
-public class Parser {
-    public String parseCommand(String userInput) {
-        return userInput.split(" ")[0].trim();
-    }
-}
