--- src/main/java/Friday/task/Todo.java
@@ -1,13 +1,0 @@
-package Friday.task;
-
-public class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
